feat(webp): 添加对动画WebP的支持并重构相关代码
将动画WebP相关功能从anim.go迁移到webp.go和capi.go
新增AnimInfo和Frame类型用于处理动画信息
实现动画检测、信息获取、帧解码等功能
